requirements.txt
websocket
python-dotenv
questionary
selenium
undetected-chromedriver
rich